requirements.txt
python-dotenv>=0.19.0
telethon>=1.24.0
opencv-python>=4.5.3.56
numpy>=1.21.2
cryptography>=3.4.7
scikit-image>=0.18.3
asyncio>=3.4.3